+++ Tests/StreamChatTests/APIClient/Endpoints/ChannelEndpoints_Tests.swift
@@ -353,6 +353,33 @@ final class ChannelEndpoints_Tests: XCTestCase {
         XCTAssertEqual(AnyEndpoint(expectedEndpoint), AnyEndpoint(endpoint))
         XCTAssertEqual("channels/\(cid.type.rawValue)/\(cid.id)", endpoint.path.value)
     }
+    
+    func test_addMembers_withHideHistoryBefore_buildsCorrectly() {
+        let cid = ChannelId.unique
+        let userIds: Set<UserId> = Set([UserId.unique])
+        let members = userIds.map { MemberInfoRequest(userId: $0, extraData: ["is_premium": true]) }
+        let hideHistoryBefore = Date()
+
+        let expectedEndpoint = Endpoint<EmptyResponse>(
+            path: .channelUpdate(cid.apiPath),
+            method: .post,
+            queryItems: nil,
+            requiresConnectionId: false,
+            body: ["add_members": AnyEncodable(members), "hide_history_before": AnyEncodable(hideHistoryBefore)]
+        )
+
+        // Build endpoint
+        let endpoint: Endpoint<EmptyResponse> = .addMembers(
+            cid: cid,
+            members: members,
+            hideHistory: true,
+            hideHistoryBefore: hideHistoryBefore
+        )
+
+        // Assert endpoint is built correctly
+        XCTAssertEqual(AnyEndpoint(expectedEndpoint), AnyEndpoint(endpoint))
+        XCTAssertEqual("channels/\(cid.type.rawValue)/\(cid.id)", endpoint.path.value)
+    }
 
     func test_removeMembers_buildsCorrectly() {
         let cid = ChannelId.unique